refactor: 프로필 API 타입 및 프로필 타입 수정
