+++ src/features/profiles/model/type.ts
@@ -1,0 +1,9 @@
+import type { Profile } from "@/entities/profiles/model/type";
+
+export type ProfileList = {
+  items: Profile[];
+  total: number;
+  page: number;
+  first: boolean;
+  last: boolean;
+};
